+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 升级
 
 * 3.11
+  * 添加 options.customWysiwygMobileToolbar
   * 添加 wavedromRender 方法
   * 添加 options.preview.markdown.sup
   * 添加 options.preview.markdown.sub
@@ -32,6 +33,8 @@
 * [open issues](https://github.com/Vanessa219/vditor/issues)
 
 ### v3.11.4 / 2026
+
+* [支持自定义移动端 WYSIWYG 工具栏 `customWysiwygMobileToolbar`](https://github.com/Vanessa219/vditor/issues/414) `引入特性`
 
 ### v3.11.3 / 2026-08-11